\name{getFamilies}
\alias{getFamilies}
\title{
   Lists all distribution families with their parameters
}
\description{
   This function returns a list of lists of all distribution families with their parameters and scans packages if desired.
   
}
\usage{
getFamilies(all.packages, file=...)

}

\arguments{
   **internal**
  \item{all.packages}{
      vector of package names to scan for distribution families or
     \code{TRUE} to scan all packages  or \code{FALSE} to scan base packages only or \code{MISSING} to return families saved in file
  }
}

\details{
  
}
\value{
  %description of output%
}

\author{
   Tim Glockner
}
\note{
  **further notes**
}

% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 **objects to See Also as \code{\link{help}}, **~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## getFamilies()
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
}
\keyword{distribution}
